bert-document-classification-reproduce2/requirements.txt
pandas 
tqdm 
scikit-learn 
pytorch-pretrained-bert
jupyter
matplotlib
numpy

# cli
fire

# author embeddings
wikimapper
smart-open

# for padding
keras
tensorflow